# src/routes/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "sobermart",
      "contact": {}
    },
    "host": "kuro.asrofur.me",
    "basePath": "/sober/api",
    "securityDefinitions": {},
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/{main}/product/1": {
        "get": {
          "summary": "GetProductsById",
          "tags": [
            "Admin"
          ],
          "operationId": "GetProductsById",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "put": {
          "summary": "UpdateProduct",
          "tags": [
            "Admin"
          ],
          "operationId": "UpdateProduct",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            },
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "delete": {
          "summary": "DeleteProduct",
          "tags": [
            "Admin"
          ],
          "operationId": "DeleteProduct",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "main",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": []
        }
      },
...
      "/": {
        "get": {
          "summary": "GetAllReview",
          "tags": [
            "Admin"
          ],
          "operationId": "GetAllReview",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "delete": {
          "summary": "DeleteReview",
          "tags": [
            "Admin"
          ],
          "operationId": "DeleteReview",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "put": {
          "summary": "UpdateShipment",
          "tags": [
            "Admin"
          ],
          "operationId": "UpdateShipment",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        },
        "post": {
          "summary": "UpdateProfile",
          "tags": [
            "User"
          ],
          "operationId": "UpdateProfile",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": false,
              "default": "Bearer {token}",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
    },
    "security": [],
...
  }